refactor: move RevisionStatusWidget to a separate file
